Another information in my favorite series
Hotel De Luna is a series about a lost soul that is existing in the real word for over thousand of years which led by actress IU. One piece is an anime series that tells the story who wants to be a King of the Pirates.  Mouse is a kdrama series about a killer that disguised as a cop. Lastly, the flower of evil is a suspense melodrama about a man who hides his cruel past and changes his identity, while his detective wife tracks down his past.